eddn.archive now appears to be working.
A lot of work here is clean ups of various mechanisms:

	1) Back-Off Sleep now actually works
	2) We get the last gatewayTimestamp per 'archive', not just
	overall.
	3) eddn.database.latestMessageTimestamp() now sorts on the
	json-embedded gatewayTimestamp.
	4) It's now the main eddn-archive function in eddnlistener.py
	that loops over archive types.

Given the possibility of longer downtimes and thus needing to back-fill
a LOT of data from eddn-archive the code needs further changes to commit
messages to the db as soon as received from eddn-archive.  Currently we
collect them all up in an array and iterate over it later.