Changes archive code to commit ASAP
Rather than committing all the data we just pulled from eddn-archive in
one go, which could be days worth for each schema type, we now commit
the messages we got in a single eddn-archive response immediately.

This has necessitated some code re-organisation to ensure that
eddn.archive functions can do the same validation as the main program
loop without duplicating the code.  Hence we now have eddn/utils.py (not
a class).

Also fixes a bug where, due to passing 'from' timestamp with ".0" on the
end, eddn-archive was returning an entire day's worth of data, not just
from the timestamp we specified.

Additionally there was a bug with the looping over schema types and
detecting the last time stamp.  Now a single loop.