Adds sleep after EDDN Archive thread start
This is attempting to avoid the case where, although we only start the
EDDN Archive thread after first valid message is received, we might
still then insert that into the database before the archive thread runs.

Also, we'll bail out of eddn.archive.getTimeRange if start and end are
the same datestamp.