Ensures eddn.archive uses correct timestamps per archive type
We were looping over archivetypes and using the latest timestamp in each
at the time we started processing the type.  If a new message had been
received for that type since we started the archive thread then:

	1) We'd miss out on all the messages we were meant to backfill!
	2) In getTimeRange() we'd have 'start' AFTER 'end'.

So instead gather all the "last seen" timestamps per archive before we
even start the archive thread (which is before the triggering first
message is stored in the database) and iterate over those in the thread.