Makes EDDN archive startup more efficient, and handles SQL connection drop
If insertMessage() gets an exception about the database connection
dropping it will now retry a number of times, with a sleep between
each.

DB Schema updated to have RDBMS 'schemaref' and 'gatewaytimestamp'
columns which are now used in the EDDN archive startup queries.  They're
near instant even with 3.9 million rows, compared to taking 'forever'
before.
  Could possibly just have added a gin index on the fields, but prefer a
proper TIMESTAMP column for the gatewaytimestamp to be sure.